+++ src/utils/spade.rs
@@ -4,7 +4,7 @@ use crate::math::Real;
 ///
 /// Returns 0.0 if the coordinate is smaller than `spade::MIN_ALLOWED_VALUE`.
 /// Returns `spade::MAX_ALLOWED_VALUE` the coordinate is larger than `spade::MAX_ALLOWED_VALUE`.
-pub fn sanitize_coord(coord: Real) -> Real {
+pub fn sanitize_spade_coord(coord: Real) -> Real {
     let abs = coord.abs();
 
     #[allow(clippy::unnecessary_cast)]
@@ -21,6 +21,7 @@ pub fn sanitize_coord(coord: Real) -> Real {
     coord
 }
 
-pub fn sanitize_point(point: spade::Point2<Real>) -> spade::Point2<Real> {
-    spade::Point2::new(sanitize_coord(point.x), sanitize_coord(point.y))
+/// Ensures the coordinates of the given point don’t go out of the bounds of spade’s acceptable values.
+pub fn sanitize_spade_point(point: spade::Point2<Real>) -> spade::Point2<Real> {
+    spade::Point2::new(sanitize_spade_coord(point.x), sanitize_spade_coord(point.y))
 }