+++ CHANGELOG.md
@@ -2,12 +2,20 @@
 
 ## Unreleased
 
+### Added
+
+- Add `Triangle::robust_scaled_normal` and `Triangle::robust_normal` as a more robust way to compute the triangles
+  normal for thin triangles that generally cause numerical instabilities.
+- Add `Triangle::angle_closest_to_90` to find the triangle’s vertex with an angle closest to 90 degree.
+- Add the `wavefront` feature that enables `TriMesh::to_obj_file` for exporting a mesh as an obj file.
+
 ### Modified
 
 - `TypedShape::Custom(u32)` is now `TypedShape::Custom(&dyn Shape)`.
 - `Real` is now exposed through a `use` statement,
   so that an indirection is removed in documentation:
   previous occurrences of `Real` now show `f32` or `f64`.
+- Significantly improved the general stability of mesh/mesh intersection calculation.
 
 ## v0.16.1